Adding some things to bash_profile

+++ .bash_profile
@@ -1,21 +1,59 @@
-#ALIASES
+# Detect OS
+OS=`lowercase \`uname\``
+if [ $OS == "darwin" ]; then
+  OS="mac"
+else
+  OS="linux"
+fi
 
+#ALIASES
 alias ..="cd .."
 alias ...="cd ../.."
 alias ....="cd ../../.."
 alias .....="cd ../../../.."
-alias ll='ls -l'
+alias ll='ls -hal'
 alias rm='rm -i'
-alias co='sh $HOME/bin/rmate'
-alias gitInfo='ssh git@git.corp.appnexus.com info'
 alias diskspace='du -S | sort -n -r | more'
-alias adnxs='cd /usr/local/adnxs'
-alias maestroui='cd /usr/local/adnxs/maestro3-ui'
-alias maestroapi='cd /usr/local/adnxs/maestro3-api'
-alias tasker='cd /usr/local/adnxs/tasker-api'
-alias eos_rm='eos ps | grep $USER | awk '"'"'{ print $1 }'"'"' | while read inst ;  do eos kill -i $inst && eos rmc -i $inst; done'
-alias refresh='eval `ssh-agent`; ssh-add'
+if [ "$OS" = "linux"]; then
+    alias co='sh $HOME/bin/rmate'
+    alias gitInfo='ssh git@git.corp.appnexus.com info'
+    alias adnxs='cd /usr/local/adnxs'
+    alias maestroui='cd /usr/local/adnxs/maestro3-ui'
+    alias maestroapi='cd /usr/local/adnxs/maestro3-api'
+    alias tasker='cd /usr/local/adnxs/tasker-api'
+    alias eos_rm='eos ps | grep $USER | awk '"'"'{ print $1 }'"'"' | while read inst ;  do eos kill -i $inst && eos rmc -i $inst; done'
+    alias refresh='eval `ssh-agent`; ssh-add'
+fi
 
+# Finding things
+function findin () {
+  find . -exec grep -q "$1" '{}' \; -print
+}
+
+# Extracting files
+extract () {
+  if [ -f $1 ] ; then
+    case $1 in
+      *.tar.bz2)   tar xvjf $1    ;;
+      *.tar.gz)    tar xvzf $1    ;;
+      *.tgz)       tar xvzf $1    ;;
+      *.bz2)       bunzip2 $1     ;;
+      *.rar)       unrar x $1     ;;
+      *.gz)        gunzip $1      ;;
+      *.tar)       tar xvf $1     ;;
+      *.tbz2)      tar xvjf $1    ;;
+      *.tgz)       tar xvzf $1    ;;
+      *.zip)       unzip $1       ;;
+      *.Z)         uncompress $1  ;;
+      *.7z)        7z x $1        ;;
+      *)           echo "'$1' cannot be extracted via >extract<" ;;
+    esac
+  else
+    echo "'$1' is not a valid file"
+  fi
+}
+
+# Print specific column
 function fawk {
     first="awk '{print "
     last="}'"
@@ -23,13 +61,40 @@ function fawk {
     eval $cmd
 }
 
+
 # Activate z command
 . $HOME/bin/z/z.sh
 
+# Exposing editor for things
+export EDITOR = 'vim'
+if [ $OS != "mac" ]; then
+  alias vi="vim"
+fi
+
 # Setting PATH for Python 2.7 AND Ruby
-# The orginal version is saved in .bash_profile.pysave
-PATH="/Users/dtrujillo/Python-2.7.3:/Users/dtrujillo/bin:/usr/local/opt/ruby/bin:/Library/Frameworks/Python.framework/Versions/2.7/bin:$HOME/bin:$HOME/git_alias:$HOME:${PATH}"
-export PATH
+export PATH="$HOME:$PATH"
+export PATH="$HOME/bin:$PATH"
+export PATH="$HOME/git_alias:$PATH"
+if [ "$OS" = "mac" ]; then
+    export PATH="$HOME/Python-2.7.3:$PATH"
+    export PATH="/usr/local/opt/ruby/bin:$PATH"
+    export PATH="/Library/Frameworks/Python.framework/Versions/2.7/bin:$PATH"
+fi
+
+export DATE=$(date +%Y-%m-%dT%H:%M:%S%z)
+
+# Customize history as we know it
+# Don't write duplicate lines in the bash_history
+export HISTCONTROL=ignoredups
+
+# Append to the history file, don't overwrite it. This will cause some
+# duplicates, even with the setting above since t a new history setting is
+# saved with each session.
+shopt -s histappend
+
+# Large command history file
+HISTFILESIZE=1000000
+HISTSIZE=10000
 
 # Setting for the new UTF-8 terminal support in Lion
 export LANGUAGE=en_US.UTF-8
@@ -37,11 +102,12 @@ export LANG=en_US.UTF-8
 export LC_ALL=en_US.UTF-8
 export SSH_AUTH_SOCK=$HOME/.ssh/ssh_auth_sock
 
-# Load bashrf if it exists
+# Load bashrc if it exists
 if [ -f $HOME/.bashrc ]; then
    source $HOME/.bashrc
 fi
 
+# Setting up PROMPT
 source /etc/bash_completion.d/git
 
 # Note these first four lines are optional
